Created new Handler | Moved handler into own file
