+++ main.go
@@ -1,15 +1,9 @@
 package main
 
 import (
-	"log"
-	"net/http"
-	"strings"
-
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-contrib/sessions/cookie"
 	"github.com/gin-gonic/gin"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"golang.org/x/crypto/bcrypt"
 )
 
 func main() {
@@ -25,149 +19,14 @@ func main() {
 	v1.Use(authRequired)
 	{
 		v1.GET("/tracks", handleGetTracks)
+		v1.GET("tracks/:name", handlerGetOneTracks)
 		v1.POST("/tracks", handleCreateTrack)
-		v1.GET("/me", me)
-		v1.GET("/status", status)
 
 	}
+
+	// currently for dev ussage
+	r.GET("/me", authRequired, me)
+	r.GET("/status", authRequired, status)
 
 	r.Run(":8080") // listen and serve on 0.0.0.0:8080
-}
-
-func handleGetTracks(c *gin.Context) {
-	var loadedTracks, err = GetAllTracks()
-	if err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"msg": err})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"tracks": loadedTracks})
-}
-
-func handleCreateTrack(c *gin.Context) {
-	var track Track
-	if err := c.ShouldBindJSON(&track); err != nil {
-		log.Print(err)
-		c.JSON(http.StatusBadRequest, gin.H{"msg": err})
-		return
-	}
-	log.Print(&track)
-	id, err := Create(&track)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"msg": err})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"_id": id})
-}
-
-func authRequired(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get("userName")
-	if user == nil {
-		// Abort the request with the appropriate error code
-		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-		return
-	}
-	// Continue down the chain to handler etc
-	c.Next()
-}
-
-// login is a handler that parses a form and checks for specific data
-func login(c *gin.Context) {
-	session := sessions.Default(c)
-	var credentials Credentials
-	if err := c.ShouldBindJSON(&credentials); err != nil {
-		log.Print(err)
-		c.JSON(http.StatusBadRequest, gin.H{"msg": err})
-		return
-	}
-
-	// Validate form input
-	if strings.Trim(credentials.Username, " ") == "" || strings.Trim(credentials.Password, " ") == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Parameters can't be empty"})
-		return
-	}
-
-	credDB, _ := GetCredentials(&credentials)
-
-	if credDB == nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Authentication failed"})
-		return
-	}
-
-	result := comparePasswords(credDB["password"].(string), []byte(credentials.Password))
-
-	// Check for username and password match, usually from a database
-	if result == false {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Authentication failed"})
-		return
-	}
-
-	// Save the username in the session
-	session.Set("userName", credDB["username"].(string))
-	session.Set("userID", credDB["_id"].(primitive.ObjectID).Hex())
-	if err := session.Save(); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"message": result})
-}
-
-func logout(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get("userName")
-	if user == nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid session token"})
-		return
-	}
-	session.Delete("userName")
-	if err := session.Save(); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"message": "Successfully logged out"})
-}
-
-func me(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get("userName")
-	userID := session.Get("userID")
-	c.JSON(http.StatusOK, gin.H{"user": user, "ID": userID})
-}
-
-func status(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{"status": "You are logged in"})
-}
-
-//temp funktion for generting pw hashes
-func pwgen(c *gin.Context) {
-	var credentials Credentials
-	if err := c.ShouldBindJSON(&credentials); err != nil {
-		log.Print(err)
-		c.JSON(http.StatusBadRequest, gin.H{"msg": err})
-		return
-	}
-	pw := hashAndSalt([]byte(credentials.Password))
-	c.JSON(http.StatusOK, gin.H{"Hash": pw})
-}
-
-func hashAndSalt(pwd []byte) string {
-
-	// Use GenerateFromPassword to hash & salt pwd
-	hash, err := bcrypt.GenerateFromPassword(pwd, bcrypt.MinCost)
-	if err != nil {
-		log.Println(err)
-	}
-	// return the hash as a string for better storage
-	return string(hash)
-}
-
-func comparePasswords(hashedPwd string, plainPwd []byte) bool {
-	byteHash := []byte(hashedPwd)
-	err := bcrypt.CompareHashAndPassword(byteHash, plainPwd)
-	if err != nil {
-		log.Println(err)
-		return false
-	}
-
-	return true
 }
